refactor: 예외처리 메서드 GlobalReservationExceptionHandler로 분리
