+++ src/main/java/roomescape/controller/GlobalReservationExceptionHandler.java
@@ -1,0 +1,19 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+public class GlobalReservationExceptionHandler {
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<Void> handleIllegalException(IllegalArgumentException e) {
+        System.out.println("IllegalArgumentException occurred: " + e.getMessage());
+        return ResponseEntity.badRequest().build();
+    }
+
+    @ExceptionHandler(IllegalStateException.class)
+    public ResponseEntity<Void> handleIllegalException(IllegalStateException e) {
+        System.out.println("IllegalStateException occurred: " + e.getMessage());
+        return ResponseEntity.internalServerError().build();
+    }
+}
