+++ src/main/java/roomescape/controller/RoomescapeController.java
@@ -44,15 +44,4 @@ public class RoomescapeController {
         return ResponseEntity.ok().build();
     }
 
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<Void> handleIllegalException(IllegalArgumentException e) {
-        System.out.println("IllegalArgumentException occurred: " + e.getMessage());
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(IllegalStateException.class)
-    public ResponseEntity<Void> handleIllegalException(IllegalStateException e) {
-        System.out.println("IllegalStateException occurred: " + e.getMessage());
-        return ResponseEntity.internalServerError().build();
-    }
 }
